LPS-49294 Move getImmutableFieldTypes and getAnnotations to BaseSourceProcessor
